+++ packages/playground/server-components/src/components/Config.client.jsx
@@ -1,0 +1,7 @@
+import {ShopifyProvider} from '@shopify/hydrogen/client';
+
+export default function ClientConfig({shopifyConfig, children}) {
+  return (
+    <ShopifyProvider shopifyConfig={shopifyConfig}>{children}</ShopifyProvider>
+  );
+}