+++ packages/hydrogen/CHANGELOG.md
@@ -7,7 +7,12 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+<<<<<<< HEAD
 - Fix index routes. See [#562](https://github.com/Shopify/hydrogen/issues/562)
+||||||| parent of 9946e7f (fix: update changelog and add e2e tests)
+=======
+- Feature - `<ShopifyProvider>` can again be used in server components
+>>>>>>> 9946e7f (fix: update changelog and add e2e tests)
 - dx: Correct Typescript issue where `as` was a default prop for all components when it should not be
 - New React hook `useScriptLoader` is available to more easily load external scripts
 - DX: Update types and docs for `useCart()` hook and `<CartProvider>`